+++ CHANGELOG.md
@@ -399,6 +399,7 @@
 - [By-type pattern matching][3742]
 - [Fix performance of method calls on polyglot arrays][3781]
 - [Missing foreign language generates proper Enso error][3798]
+- [Made Vector performance to be on par with Array][3811]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -453,6 +454,7 @@
 [3742]: https://github.com/enso-org/enso/pull/3742
 [3781]: https://github.com/enso-org/enso/pull/3781
 [3798]: https://github.com/enso-org/enso/pull/3798
+[3811]: https://github.com/enso-org/enso/pull/3811
 
 # Enso 2.0.0-alpha.18 (2021-10-12)